portrait: draw a likeness - small line-art model, sub-degree servo commands
Two changes, either of which alone leaves the drawing unrecognisable.

Line art. Canny and DoG trace whatever has contrast, so on a real photo they
find the hair silhouette and the glasses and miss the eyes, nose and mouth --
a recognisable haircut wearing spectacles, with no face in it. The portrait
step now prefers a small Informative Drawings generator run on the CPU through
onnxruntime: 17 MB, ~4.5 M parameters, about a second a frame, no GPU, so the
same path runs on the UNO Q. GrabCut stays, but only as a mask - the model
would otherwise faithfully draw the building behind the subject. Missing model
or missing onnxruntime falls straight back to the old tracer, so CI and a
fresh clone keep working with nothing downloaded.

Servo resolution. This was the bigger one. The paper sits ~175 mm from the base
axis, so one degree of base rotation sweeps the pen 3.05 mm and a 40 mm sheet
spans 13 degrees: at whole degrees the drawing has about a dozen addressable
columns and comes out as rubble. The M protocol now carries fractional degrees
end to end - kinematics rounds to motion.servo_decimals (default 1) instead of
casting to int, arm_client formats with %g, and the simulator, the Gazebo
bridge and the UNO Q agent all parse with float(). On hardware the sketch
writes writeMicroseconds() rather than write(), which takes whole degrees and
would discard the fraction at the last moment. servo_decimals: 0 restores the
old behaviour for firmware that cannot take fractions.

Same plan, same 1111 points, whole degrees vs tenths: docs/images/servo-resolution.png.

Also:
- planner join_gap_mm 4.0 -> 0.8. At 4 mm on a 40 mm sheet the pen stayed down
  across a tenth of the page and joined unrelated strokes with straight lines.
- vectorize keeps more detail (min_contour_len 24 -> 14, epsilon 2.0 -> 1.0,
  cap 140 -> 240) now that there is real detail worth keeping.
- expose the Gazebo bridge's arrival tolerance and tighten it to 0.15 deg
  (~0.45 mm); at 0.75 deg it answered OK while 2.3 mm from the target.
- tests for fractional angles, the servo_decimals: 0 opt-out, and that 0.1 deg
  is a sub-millimetre step at the paper.
- README gains a "Getting a likeness" section covering both failure modes.

Known issue, documented in sim/gazebo/README.md: the Gazebo trace still does
not match the plan, though the software simulator and a direct plot of the
planned moves both do. Servo quantisation, stroke segmentation, pen-up sag and
the arrival tolerance have each been ruled out with evidence. Relatedly, the
end-to-end check scores each sample against the nearest planned segment and so
passes a visibly wrong drawing; it needs an in-order metric.

+++ app_lab/braccio_remote_agent/python/main.py
@@ -38,7 +38,10 @@ def handle_command(command):
         return "ERR"
 
     try:
-        values = [int(value) for value in parts[1:]]
+        # float(): the sketchbot streams tenths of a degree, because a whole
+        # degree moves the pen ~3 mm at the paper. Whole-degree senders still
+        # parse fine.
+        values = [float(value) for value in parts[1:]]
     except ValueError:
         return "ERR"
